requirements.txt
google-api-python-client
youtube-transcript-api
streamlit
openai
python-dotenv
nltk